updating label type

+++ train_refactor.py
@@ -62,7 +62,7 @@ def decode(serialized_example):
     # length mnist.IMAGE_PIXELS) to a uint8 tensor with shape
     # [mnist.IMAGE_PIXELS].
     data = tf.decode_raw(features['pointclouds_pl'], tf.float32)
-    labels = tf.decode_raw(features['labels_pl'], tf.float32)
+    labels = tf.decode_raw(features['labels_pl'], tf.int32)
     sample_weights = tf.decode_raw(features['smpws_pl'], tf.float32)
     # data.set_shape((mnist.IMAGE_PIXELS))
